tests/test_explainability.py
```python
# Importing necessary libraries
import unittest
import numpy as np
from src.explainability import Explainability

class TestExplainability(unittest.TestCase):
    def setUp(self):
        """
        Set up a dummy model and data for testing
        """
        self.model = DummyModel()
        self.data = np.random.rand(10, 5)
        self.explainability = Explainability(self.model)

    def test_shap_explain(self):
        """
        Test SHAP explainability
        """
        shap_values = self.explainability.shap_explain(self.data)
        self.assertIsNotNone(shap_values)

    def test_lime_explain(self):
        """
        Test LIME explainability
        """
        lime_values = self.explainability.lime_explain(self.data)
        self.assertIsNotNone(lime_values)

    def test_captum_explain(self):
        """
        Test Captum explainability
        """
        captum_values = self.explainability.captum_explain(self.data)
        self.assertIsNotNone(captum_values)

    def test_interpret_explain(self):
        """
        Test InterpretML explainability
        """
        interpret_values = self.explainability.interpret_explain(self.data)
        self.assertIsNotNone(interpret_values)


class DummyModel:
    """
    A dummy model for testing purposes
    """
    def __init__(self):
        self.layer1 = DummyLayer()

    def predict(self, data):
        return np.random.rand(data.shape[0], 1)


class DummyLayer:
    """
    A dummy layer for testing purposes
    """
    pass


if __name__ == '__main__':
    unittest.main()
```
